menu and bug fixes in snake_game.c
I saw that project when I was trying to learn how to make this game on linux. Its a pretty cool project and gave me the right to change some little things in it. First I transcript most of the variables to my mother language (portuguese) just to understand better what is happening. Later I made a main menu for the player, where they can start, see controls and exit if he want it. Then, I fixed a bug happening when the snake turn in a size above of 30, the fruits was starting to move and the game was just crashing at size 33. Finally a make the fruits stop spawning  inside snake's body and made some minor changes just because I want it. If you did not understand something, please contact me :).